+++ primefaces/src/main/java/org/primefaces/component/treetable/TreeTableBase.java
@@ -41,6 +41,8 @@ public abstract class TreeTableBase extends UITree implements Widget, ClientBeha
     public static final String COMPONENT_FAMILY = "org.primefaces.component";
 
     public static final String DEFAULT_RENDERER = "org.primefaces.component.TreeTableRenderer";
+    public static final String FILTER_PRUNE_NONE = "none";
+    public static final String FILTER_PRUNE_DESCENDANTS = "descendants";
 
     public enum PropertyKeys {
 
@@ -77,6 +79,7 @@ public abstract class TreeTableBase extends UITree implements Widget, ClientBeha
         rows,
         first,
         filterBy,
+        filterPrune,
         filterNormalize,
         globalFilter,
         globalFilterFunction,
@@ -411,6 +414,18 @@ public abstract class TreeTableBase extends UITree implements Widget, ClientBeha
 
     public void setFilterDelay(int filterDelay) {
         getStateHelper().put(PropertyKeys.filterDelay, filterDelay);
+    }
+
+    public String getFilterPrune() {
+        return (String) getStateHelper().eval(PropertyKeys.filterPrune, FILTER_PRUNE_NONE);
+    }
+
+    public void setFilterPrune(String filterPrune) {
+        getStateHelper().put(PropertyKeys.filterPrune, filterPrune);
+    }
+
+    public boolean isFilterPruneDescendants() {
+        return FILTER_PRUNE_DESCENDANTS.equals(getFilterPrune());
     }
 
     public String getCellEditMode() {